Add more values to setup.xml
The Echo finds us! And she sends the commands directly to us!
Sometimes it's nice to be wrong :)

The only problem now is that multiple devices are ignored, and
indeed if I stop advertising the one she accepted and do another
search, the new one is not found, and the old one remains.
Amazon is probably indexing by something that we're not providing
unique values for. Hopefully it's not IP or MAC address....