Add compatibility with the homee WeMo integration (#7)
* Add eventservice.xml

This file is used to determine the capabilities of the device. Because we can only turn the emulated device on or off we add the get and set binary state actions.

* Fix initial binary state

Use the with 0 initialized binary state instead it being undefined.

* Return SetBinaryStateResponse

Return a correct SetBinaryStateResponse after setting the binary state.